Ensure the cluster scaler jumps to min_capacity
Fixes #979

We currently constrain to a `MAX_CLUSTER_DELTA` when scaling up or down.
This prevents sudden large changes when the autoscaler has made a
mistake. However, an operator increasing `min_capacity` on a resource
should trump this situation. This commit makes it so that we never set a
new capacity that is lower than the current `min_capacity`.